requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
channels==4.0.0
charset-normalizer==3.3.2
Django==5.0.1
django-nextjs==2.4.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
websockets==12.0
yarl==1.9.4